requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # For Excel file support
python-dotenv>=1.0.0
openai>=1.12.0  # For Azure OpenAI integration
azure-core>=1.29.5
typing-extensions>=4.7.0
python-json-logger>=2.0.7
